+++ src/Extension/CommonMark/Parser/Inline/BacktickParser.php
@@ -19,25 +19,21 @@ namespace League\CommonMark\Extension\CommonMark\Parser\Inline;
 use League\CommonMark\Extension\CommonMark\Node\Inline\Code;
 use League\CommonMark\Node\Inline\Text;
 use League\CommonMark\Parser\Inline\InlineParserInterface;
+use League\CommonMark\Parser\Inline\InlineParserMatch;
 use League\CommonMark\Parser\InlineParserContext;
 
 final class BacktickParser implements InlineParserInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getCharacters(): array
+    public function getMatchDefinition(): InlineParserMatch
     {
-        return ['`'];
+        return InlineParserMatch::regex('`+');
     }
 
-    public function parse(InlineParserContext $inlineContext): bool
+    public function parse(string $match, InlineParserContext $inlineContext): bool
     {
+        $ticks  = $match;
         $cursor = $inlineContext->getCursor();
-        $ticks  = $cursor->match('/^`+/');
-        if ($ticks === null) {
-            return false; // This should never happen
-        }
+        $cursor->advanceBy(\mb_strlen($ticks));
 
         $currentPosition = $cursor->getPosition();
         $previousState   = $cursor->saveState();