+++ src/Extension/SmartPunct/PunctuationParser.php
@@ -18,32 +18,29 @@ namespace League\CommonMark\Extension\SmartPunct;
 
 use League\CommonMark\Node\Inline\Text;
 use League\CommonMark\Parser\Inline\InlineParserInterface;
+use League\CommonMark\Parser\Inline\InlineParserMatch;
 use League\CommonMark\Parser\InlineParserContext;
 
 final class PunctuationParser implements InlineParserInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getCharacters(): array
+    public function getMatchDefinition(): InlineParserMatch
     {
-        return ['-', '.'];
+        return InlineParserMatch::oneOf('-', '.');
     }
 
-    public function parse(InlineParserContext $inlineContext): bool
+    public function parse(string $match, InlineParserContext $inlineContext): bool
     {
         $cursor = $inlineContext->getCursor();
-        $ch     = $cursor->getCharacter();
 
         // Ellipses
-        if ($ch === '.' && $matched = $cursor->match('/^\\.( ?\\.)\\1/')) {
+        if ($match === '.' && $matched = $cursor->match('/^\\.( ?\\.)\\1/')) {
             $inlineContext->getContainer()->appendChild(new Text('…'));
 
             return true;
         }
 
         // Em/En-dashes
-        if ($ch === '-' && $matched = $cursor->match('/^(?<!-)(-{2,})/')) {
+        if ($match === '-' && $matched = $cursor->match('/^(?<!-)(-{2,})/')) {
             $count   = \strlen($matched);
             $enDash  = '–';
             $enCount = 0;