+++ src/Extension/SmartPunct/QuoteParser.php
@@ -18,6 +18,7 @@ namespace League\CommonMark\Extension\SmartPunct;
 
 use League\CommonMark\Delimiter\Delimiter;
 use League\CommonMark\Parser\Inline\InlineParserInterface;
+use League\CommonMark\Parser\Inline\InlineParserMatch;
 use League\CommonMark\Parser\InlineParserContext;
 use League\CommonMark\Util\RegexHelper;
 
@@ -26,27 +27,19 @@ final class QuoteParser implements InlineParserInterface
     public const DOUBLE_QUOTES = [Quote::DOUBLE_QUOTE, Quote::DOUBLE_QUOTE_OPENER, Quote::DOUBLE_QUOTE_CLOSER];
     public const SINGLE_QUOTES = [Quote::SINGLE_QUOTE, Quote::SINGLE_QUOTE_OPENER, Quote::SINGLE_QUOTE_CLOSER];
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getCharacters(): array
+    public function getMatchDefinition(): InlineParserMatch
     {
-        return \array_merge(self::DOUBLE_QUOTES, self::SINGLE_QUOTES);
+        return InlineParserMatch::oneOf(...\array_merge(self::DOUBLE_QUOTES, self::SINGLE_QUOTES));
     }
 
     /**
      * Normalizes any quote characters found and manually adds them to the delimiter stack
      */
-    public function parse(InlineParserContext $inlineContext): bool
+    public function parse(string $match, InlineParserContext $inlineContext): bool
     {
-        $cursor    = $inlineContext->getCursor();
-        $character = $cursor->getCharacter();
+        $cursor = $inlineContext->getCursor();
 
-        if ($character === null) {
-            return false;
-        }
-
-        $normalizedCharacter = $this->getNormalizedQuoteCharacter($character);
+        $normalizedCharacter = $this->getNormalizedQuoteCharacter($match);
 
         $charBefore = $cursor->peek(-1);
         if ($charBefore === null) {