+++ src/Extension/TaskList/TaskListItemMarkerParser.php
@@ -16,19 +16,17 @@ namespace League\CommonMark\Extension\TaskList;
 use League\CommonMark\Extension\CommonMark\Node\Block\ListItem;
 use League\CommonMark\Node\Block\Paragraph;
 use League\CommonMark\Parser\Inline\InlineParserInterface;
+use League\CommonMark\Parser\Inline\InlineParserMatch;
 use League\CommonMark\Parser\InlineParserContext;
 
 final class TaskListItemMarkerParser implements InlineParserInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getCharacters(): array
+    public function getMatchDefinition(): InlineParserMatch
     {
-        return ['['];
+        return InlineParserMatch::oneOf('[ ]', '[x]');
     }
 
-    public function parse(InlineParserContext $inlineContext): bool
+    public function parse(string $match, InlineParserContext $inlineContext): bool
     {
         $container = $inlineContext->getContainer();
 
@@ -40,10 +38,7 @@ final class TaskListItemMarkerParser implements InlineParserInterface
         $cursor   = $inlineContext->getCursor();
         $oldState = $cursor->saveState();
 
-        $m = $cursor->match('/\[[ xX]\]/');
-        if ($m === null) {
-            return false;
-        }
+        $cursor->advanceBy(3);
 
         if ($cursor->getNextNonSpaceCharacter() === null) {
             $cursor->restoreState($oldState);
@@ -51,7 +46,7 @@ final class TaskListItemMarkerParser implements InlineParserInterface
             return false;
         }
 
-        $isChecked = $m !== '[ ]';
+        $isChecked = $match !== '[ ]';
 
         $container->appendChild(new TaskListItemMarker($isChecked));