+++ src/Parser/Inline/NewlineParser.php
@@ -22,15 +22,12 @@ use League\CommonMark\Parser\InlineParserContext;
 
 final class NewlineParser implements InlineParserInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function getCharacters(): array
+    public function getMatchDefinition(): InlineParserMatch
     {
-        return ["\n"];
+        return InlineParserMatch::regex('\\n');
     }
 
-    public function parse(InlineParserContext $inlineContext): bool
+    public function parse(string $match, InlineParserContext $inlineContext): bool
     {
         $inlineContext->getCursor()->advanceBy(1);